Wrap s3-sftp code in Lambda function
Lightweight Lambda wrapper in the handler function that does the Lambda specific stuff and calls the main logic.

Optional extra import path in vendored/ simplifying the package build.
Have to go fetch a pre-compiled cryptography module from S3 for now because it needs to be compiled on the distro it's running on (Amazon Linux).