First pass at Terraform module for deploying the S3-SFTP bridge
Not overly happy with the static usage example right now, could do with a bit of work to tidy up what needs configuring.
Want to add a dynamic option with a Jenkinsfile that generates the right TF configuration etc.

Might be better off taking the SFTP keys bucket out of the Terraform and just have that managed out of band to simplify things. Still want the Lambda to only have the ability to decrypt that specific KMS key though.