Add SQS for failed events
Was originally going to write some code to have the Lambda function move the message on to the queue when failing but with dead letter queues getting first class support in Terraform instead we can just configure the Lambda function to write to the SQS queue on failure automatically.
For now this needs to be done by hand by editing the Lambda function's advanced configuration to set the SQS queue as the target for dead letter handling.
In the near future I'll change the Terraform to add the dead letter configuration to the Lambda function once TF 0.9 is released (was merged in last night just as I was starting on a pull request for it).